Fix nuke and nukev functions
I wrote those two functions when I was not very familiar with
Powershell and I didn't really know how to forward the argument of a
functions to a command invoked inside of its body.
I learned about @Args so I added it. Those two functions should now be
more robust.