Improve the PSReadLine block
This patch makes sure the PSReadLine module is loaded only when the
host is ConsoleHost - as per the PSReadLine README.md on GitHub - and
also makes sure the cursor goes to the end of the line when searching
through the history.

Moreover, Ctrl+w is now bound to BackwardKillWord so that, when
deleting words, the WordDelimiters property is respected.